+++ R/progression-table-perc-drop.R
@@ -1,9 +1,5 @@
 #' @describeIn progression_table Perc Drop progression table (see Strength Training Manual)
 #' @export
-#' @references
-#'     Jovanović M. 2020. Strength Training Manual: The Agile Periodization Approach.
-#'     Independently published.
-#'
 #' @examples
 #' # ------------------------------------------
 #' # Progression Perc Drop