+++ NEWS.md
@@ -2,6 +2,11 @@
 
 * Added `conservative` option when generating progression tables, which is equal to `mfactor = 3`
 * In function `plot_progression_table()`, renamed `adjustment_multiplier` to `multiplier`, due to R "error" of assuming user-provided `adjustment` parameter to be `adjustment_multiplier`.
+* Added utils function `sig_pad()` which is used in `plot_progression_table()` to maintain number of characters and same label size.
+* Reversed the order of panels in `plot_progression_table()` from intensive-normal-extensive to extensive-normal-intensive.
+* Added `progression_variable_DI()` function to allow user to specify various progression and volume adjustment based on the reps done.
+* Added `progression_variable_RIR()` function to allow user to specify various progression and volume adjustment based on the reps done.
+* Change default `reps` parameter in  from `1:12` to `1:5`  in `generate_progression_table()` function to make plotting with `plot_progression_table()` functions visible by default due to newly added `conservative` progression.
 * Updated READMER.Rmd file
 
 # STMr 0.1.6